requirements.txt
h5py==2.10.0
matplotlib==3.1.3
numpy==1.18.1
Pillow==7.0.0
PyMCubes==0.0.9
tensorboardX==1.7
torch==1.2.0
tqdm==4.42.1
trimesh==3.2.19
